requirements.txt
python-dotenv
aiogram
openai
Deepseek